changes 2nd review
